+++ robot/robot.go
@@ -14,6 +14,7 @@ import (
 	"github.com/robocorp/rcc/common"
 	"github.com/robocorp/rcc/conda"
 	"github.com/robocorp/rcc/pathlib"
+	"github.com/robocorp/rcc/pretty"
 
 	"github.com/google/shlex"
 	"gopkg.in/yaml.v2"
@@ -54,6 +55,7 @@ type Task interface {
 
 type robot struct {
 	Tasks        map[string]*task `yaml:"tasks"`
+	Devtasks     map[string]*task `yaml:"devTasks"`
 	Conda        string           `yaml:"condaConfigFile,omitempty"`
 	PreRun       []string         `yaml:"preRunScripts,omitempty"`
 	Environments []string         `yaml:"environmentConfigs,omitempty"`
@@ -71,8 +73,22 @@ type task struct {
 	robot   *robot
 }
 
+func (it *robot) taskMap() map[string]*task {
+	if common.DeveloperFlag {
+		pretty.Note("Operating in developer mode. Using 'devTasks:' instead of 'tasks:'.")
+		return it.Devtasks
+	} else {
+		return it.Tasks
+	}
+}
+
 func (it *robot) relink() {
 	for _, task := range it.Tasks {
+		if task != nil {
+			task.robot = it
+		}
+	}
+	for _, task := range it.Devtasks {
 		if task != nil {
 			task.robot = it
 		}
@@ -281,8 +297,9 @@ func (it *robot) IgnoreFiles() []string {
 }
 
 func (it *robot) AvailableTasks() []string {
-	result := make([]string, 0, len(it.Tasks))
-	for name, _ := range it.Tasks {
+	tasks := it.taskMap()
+	result := make([]string, 0, len(tasks))
+	for name, _ := range tasks {
 		result = append(result, fmt.Sprintf("%q", name))
 	}
 	sort.Strings(result)
@@ -290,11 +307,12 @@ func (it *robot) AvailableTasks() []string {
 }
 
 func (it *robot) DefaultTask() Task {
-	if len(it.Tasks) != 1 {
+	tasks := it.taskMap()
+	if len(tasks) != 1 {
 		return nil
 	}
 	var result *task
-	for _, value := range it.Tasks {
+	for _, value := range tasks {
 		result = value
 		break
 	}
@@ -305,13 +323,14 @@ func (it *robot) TaskByName(name string) Task {
 	if len(name) == 0 {
 		return it.DefaultTask()
 	}
+	tasks := it.taskMap()
 	key := strings.TrimSpace(name)
-	found, ok := it.Tasks[key]
+	found, ok := tasks[key]
 	if ok {
 		return found
 	}
 	caseless := strings.ToLower(key)
-	for name, value := range it.Tasks {
+	for name, value := range tasks {
 		if caseless == strings.ToLower(strings.TrimSpace(name)) {
 			return value
 		}